Set execlists to be default for BSW
According to Ben, BSW hardware is not being validated for any use case
beyond execlists.  The default is to be on, and it needs to be on.

I verified that execlists do not increase test instability.  It does
appear that BSW cannot reliably pass piglit unless the tests are run
serially.