slave/grub: Always turn on preliminary hardware support
After talking to Ben it makes more sense to just have this on
universally.

This only has any affect on very new hardware, and we always end up
turning it on for those platforms with an override, it's just adding
work for no gain.

Further, it really simplifies the salt setup.