master: Add jenkins configuration
This is modified from the original server, it uses a systemd unit
instead of the original sysv init. This allows the configuration file to
be vastly simplified, since many of the arguments don't have any meaning
for the systemd setup, it also allows restart, and should allow making
dependencies for jenkins easier to manage.